+++ germanium/impl/_is_displayed_filter.py
@@ -1,0 +1,6 @@
+
+def _is_displayed_filter(item):
+    try:
+        return item.is_displayed()
+    except Exception:
+        return False