+++ germanium/util/__init__.py
@@ -7,5 +7,6 @@ from .select import select_g
 from .deselect import deselect_g
 from .get_value import get_value_g
 from .get_text import get_text_g
+from .highlight import highlight_g
 
 from germanium.impl import wait